+++ actix-http/CHANGES.md
@@ -4,8 +4,10 @@
 
 - Minimum supported Rust version (MSRV) is now 1.88.
 - Fix truncated body ending without error when connection closed abnormally. [#3067]
+- Add config/method for `TCP_NODELAY`. [#3918]
 
 [#3067]: https://github.com/actix/actix-web/pull/3067
+[#3918]: https://github.com/actix/actix-web/pull/3918
 
 ## 3.11.2